+++ pants.ini
@@ -3,5 +3,5 @@ pants_version: 1.0.1
 
 [python-repos]
 repos: [
-    "file:///3rdparty/repo/",
+    "file://%(buildroot)s/3rdparty/repo/",
   ]
