--- src/python/BUILD
@@ -1,7 +1,0 @@
-python_binary(name='bin',
-  dependencies=[
-    'src/python/lib',
-  ],
-  platforms = ['linux-x86_64', 'macosx-10.4-x86_64'],
-  source='main.py',
-)
