+++ src/python/grepg/lib/cheat_sheets_model.py
